requirements.txt
flask
python-dotenv
flask-restful
flask-sqlalchemy
flask-marshmallow
sqlalchemy-filters
Faker
colorama